Move test to dedicated issue spec
The test for #9 is quite specific so it's moved to a dedicated class to
not confuse readers of PluginDescriptorGenerationFuncTest.
